requirements.txt
langchain
langchain-community
pypdf
streamlit
faiss-cpu
boto3
python-dotenv
